refactor: simplify converter
